compelete project

+++ myapp/views.py
@@ -2,11 +2,15 @@ from django.shortcuts import render,redirect,HttpResponse
 from django.conf import settings
 from django.core.mail import send_mail
 from.models import*
+from django.db.models import Q
+
 
 # Create your views here.
 def indexx(request):
+    if 'email' in request.session:
         uid = User.objects.get(email=request.session['email'])
         return render(request,'index.html',{'uid':uid})
+    return redirect('login')
 
 def signupp(request):
     return render(request,'sign-up.html')
@@ -15,16 +19,18 @@ def logoutt(request):
     del request.session['email']
     return redirect('login')
 
-def ind(request):
-    return render(request,'index1.html')
-
 def index_doc(request):
-    uid = User.objects.get(email = request.session['email'])
-    return render(request,'index-doctor.html',{'uid':uid})
+    if 'email' in request.session:
+     uid = User.objects.get(email = request.session['email'])
+     return render(request,'index-doctor.html',{'uid':uid})
+
+    return redirect('login')
 
 def index_patiences(request):
-    uid = User.objects.get(email = request.session['email'])
-    return render(request,'index-patience.html',{'uid':uid})
+    if 'email' in request.session:
+        uid = User.objects.get(email = request.session['email'])
+        return render(request,'index-patience.html',{'uid':uid})
+    return redirect('login')
 
 def logins(request):
     if request.method == 'POST':
@@ -54,12 +60,13 @@ def logins(request):
     return render(request,'login.html') 
 
 def doctor_create(request):
-    if request.method == "POST":
-        try:
-            User.objects.get(email=request.POST['email'])
-            msg = "Doctor Email is already Exits "
-            return render(request,'create_doctor.html',{'msg':msg})
-        except:
+    if 'email' in request.session:
+        if request.method == "POST":
+            try:
+                User.objects.get(email=request.POST['email'])
+                msg = "Doctor Email is already Exits "
+                return render(request,'create_doctor.html',{'msg':msg})
+            except:
                 User.objects.create(
                 clinic_name=request.POST['clinic_name'],
                 gender = request.POST['gender'],
@@ -70,49 +77,62 @@ def doctor_create(request):
                 password = request.POST['password'],
             )
                 return render(request,'create_doctor.html',{'msg':'Doctor Profile Create Successfully!!'})
-    else:
-        return render(request,'create_doctor.html')
+        else:
+            return render(request,'create_doctor.html')
+    return redirect('login')
 
 
 def view_doctor(request):
-    a=User.objects.all().filter(roles ='doctor')
-    return render(request,'list_doctor.html',{'a':a})
+    if 'email' in request.session:
+        a=User.objects.all().filter(roles ='doctor')
+        return render(request,'list_doctor.html',{'a':a})
+    return redirect('login')
 
 def updates_doctor(request,pk):
-    uid = User.objects.get(id=pk)
-    if request.method == 'POST':
-        uid.name=request.POST['name']
-        uid.email=request.POST['email']
-        uid.password=request.POST['password']
-        uid.save() 
-    return render(request,'update_doctor.html',{'uid':uid,'msg':'Profile Updated Successfully!!'})
+    if 'email' in request.session:
+        uid = User.objects.get(id=pk)
+        if request.method == 'POST':
+            uid.name=request.POST['name']
+            uid.email=request.POST['email']
+            uid.password=request.POST['password']
+            uid.save() 
+            return render(request,'update_doctor.html',{'uid':uid,'msg':'Profile Updated Successfully!!'})
+        else:
+            return render(request,'update_doctor.html',{'uid':uid})
+    return redirect('login')
    
 
 
 
 def deletes_doctor(request,pk):
-    uid = User.objects.get(id=pk)
-    uid.delete()
-    return redirect('list_doctor')
+    if 'email' in request.session:
+        uid = User.objects.get(id=pk)
+        uid.delete()
+        return redirect('list_doctor')
+    return redirect('login')
 
 def updates_profile(request):
-    uid = User.objects.get(email=request.session['email'])
-    if request.method == 'POST':
-        uid.name=request.POST['name']
-        uid.email=request.POST['email']
-        uid.password=request.POST['password']
-        uid.gender=request.POST['gender']
-        uid.speciality=request.POST['speciality']
-        uid.clinic_name=request.POST['clinic_name']
-        uid.save()
-        return redirect('index_doc')
-    else:
-        return render(request,'update-profile.html',{'uid':uid})
+    if 'email' in request.session:
+        uid = User.objects.get(email=request.session['email'])
+        if request.method == 'POST':
+            uid.name=request.POST['name']
+            uid.email=request.POST['email']
+            uid.password=request.POST['password']
+            uid.gender=request.POST['gender']
+            uid.speciality=request.POST['speciality']
+            uid.clinic_name=request.POST['clinic_name']
+            uid.save()
+            return redirect('index_doc')
+        else:
+            return render(request,'update-profile.html',{'uid':uid})
+    return redirect('login')
 
 def appointment_views(request):
-    uid=User.objects.get(email=request.session['email']) 
-    many=Appointment.objects.all()
-    return render(request,'appointment-view.html',{'many':many})
+    if 'email' in request.session:
+        uid=User.objects.get(email=request.session['email']) 
+        many=Appointment.objects.all().filter(user=uid)
+        return render(request,'appointment-view.html',{'many':many})
+    return redirect('login')
     
    
 
@@ -122,132 +142,162 @@ def appointment_views(request):
 
 
 def patiences_create(request):
-    if request.method == "POST":
-        try:
-            User.objects.get(email=request.POST['email'])
-            msg = "Patience Email is already Exits "
-            return render(request,'create_patience.html',{'msg':msg})
-        except:
+    if 'email' in request.session:
+        if request.method == "POST":
+            try:
+                User.objects.get(email=request.POST['email'])
+                msg = "Patience Email is already Exits "
+                return render(request,'create_patience.html',{'msg':msg})
+            except:
                 User.objects.create(
                 address=request.POST['address'],
                 roles = request.POST['roles'],
                 name = request.POST['name'],
                 email = request.POST['email'],
                 password = request.POST['password'],
+                # gender  = request.POST['gender'],
+                profile = request.FILES['profile']
             )
                 return render(request,'create_patience.html',{'msg':'Patience Create Successfully!!'})
-    else:
-        return render(request,'create_patience.html')
+        else:
+            return render(request,'create_patience.html')
+    return redirect('login')
 
 
 def view_patience(request):
-    many=User.objects.all().filter(roles='patiences')
-    return render(request,'list_patience.html',{'many':many}) 
+    if 'email' in request.session:
+        many=User.objects.all().filter(roles='patiences')
+        return render(request,'list_patience.html',{'many':many})
+    return redirect('login') 
 
 def deletes_patience(request,pk):
-    uid=User.objects.get(id=pk)
-    uid.delete()
-    return redirect('list_patience')
+    if 'email' in request.session:
+        uid=User.objects.get(id=pk)
+        uid.delete()
+        return redirect('list_patience')
+    return redirect('login')
+
 
 def updates_patience(request,pk):
-    uid = User.objects.get(id=pk)
-    if request.method == 'POST':
-        uid.name=request.POST['name']
-        uid.email=request.POST['email']
-        uid.password=request.POST['password']
-        uid.save()
-        return redirect('list_patience')
-    else:
-        return render(request,'update_patience.html',{'uid':uid})
+    if 'email' in request.session:
+        uid = User.objects.get(id=pk)
+        if request.method == 'POST':
+            uid.name=request.POST['name']
+            uid.email=request.POST['email']
+            uid.password=request.POST['password']
+            uid.save()
+            return redirect('list_patience')
+        else:
+            return render(request,'update_patience.html',{'uid':uid})
+    return redirect('login')
 
 def appointment_status(request):
-    uid=User.objects.get(email=request.session['email'])
-    many = Appointment.objects.filter(name=uid)
-    print(many)
-    return render(request,'appointment-status.html',{'many':many})
+    if 'email' in request.session:
+        uid=User.objects.get(email=request.session['email'])
+        many = Appointment.objects.filter(Q(status=1) or Q(status=2) or Q(status=3))
+
+        print(many)
+        return render(request,'appointment-status.html',{'many':many,'uid':uid})
+    return redirect('login')
 
 
 
 
 
 def slot_add(request):
-    doctor=User.objects.get(email=request.session['email'])
-    uid=Slot.objects.all()
-    if request.method=='POST':
+    if 'email' in request.session:
+        doctor=User.objects.get(email=request.session['email'])
+        uid=Slot.objects.filter(doctor_name=doctor)
+        if request.method=='POST':
          
-        Slot.objects.create(
-        doctor_name = doctor,
-        weeks = request.POST['weeks'],
-        timeslot = request.POST['timeslot'],
-    )
-        return render(request,'slot.html',{'msg':'Slot Added Successfully!'})
-    else:
-    
-        return render(request,'slot.html',{'uid':uid})
+            Slot.objects.create(
+            doctor_name = doctor,
+            weeks = request.POST['weeks'],
+            timeslot = request.POST['timeslot'],
+            )
+            return render(request,'slot.html',{'msg':'Slot Added Successfully!'})
+        else:
+            return render(request,'slot.html',{'uid':uid})
+    return redirect('login')
 
 def view_slots(request):
-    doc = User.objects.get(email=request.session['email'])
-    many=Slot.objects.all().order_by('weeks')
-    return render(request,'view-slot.html',{'many':many})
+    if 'email' in request.session:
+        doc = User.objects.get(email=request.session['email'])
+        many = Slot.objects.all().order_by('weeks').filter(doctor_name=doc)
+        return render(request,'view-slot.html',{'many':many})
+    return redirect('login')
 
 def update_slots(request,pk):
-    id=Slot.objects.get(id=pk)
-    if request.method == 'POST':
-        id.weeks=request.POST['weeks']
-        id.timeslot=request.POST['timeslot']
-        id.save()
-        return redirect('view-slot')
-    else:
-        return render(request,'update-slot.html',{'id':id})
+    if 'email' in request.session:
+        id=Slot.objects.get(id=pk)
+        if request.method == 'POST':
+            id.weeks=request.POST['weeks']
+            id.timeslot=request.POST['timeslot']
+            id.save()
+            return redirect('view-slot')
+        else:
+            return render(request,'update-slot.html',{'id':id})
+    return redirect('login')
 
 def delete_slots(request,pk):
-    id=Slot.objects.get(id=pk)
-    id.delete()
-    return redirect('view-slot')
+    if 'email' in request.session:
+        id=Slot.objects.get(id=pk)
+        id.delete()
+        return redirect('view-slot')
+    return redirect('login')
 
 
 
 
 def profile_patience(request):
-    uid=User.objects.get(email=request.session['email'])
-    if request.method == 'POST':
-        uid.name=request.POST['name']
-        uid.email=request.POST['email']
-        uid.password=request.POST['password']
-        uid.gender=request.POST['gender']
-        uid.save()
-        return redirect('index_patience')
-    else:
-        return render(request,'patience-edit.html',{'uid':uid})
+    if 'email' in request.session:
+        uid=User.objects.get(email=request.session['email'])
+        if request.method == 'POST':
+            uid.name=request.POST['name']
+            # uid.email=request.POST['email']
+            uid.password=request.POST['password']
+            uid.gender=request.POST['gender']
+            uid.profile=request.FILES['profile']
+            uid.save()
+            return redirect('index_patience')
+        else:
+            return render(request,'patience-edit.html',{'uid':uid})
+    return redirect('login')
 
 def patience_view(request):
-    many=User.objects.all().filter(roles='doctor')[::-1]
-    return render(request,'patience-view.html',{'many':many})
+    if 'email' in request.session:
+        many=User.objects.all().filter(roles='doctor')[::-1]
+        return render(request,'patience-view.html',{'many':many})
+    return redirect('login')
 
 def appointments(request,pk):
-    uid=User.objects.get(email=request.session['email'])
-    data=Slot.objects.all()
-    return render(request,'appointment.html',{'data':data})
+    if 'email' in request.session:
+        uid=User.objects.get(email=request.session['email'])
+        data=Slot.objects.all().order_by('weeks').filter()
+        return render(request,'appointment.html',{'data':data})
+    return redirect('login')
 
 def appointments_form(request,pk):
-    user=User.objects.get(email=request.session['email'])
-    data=Slot.objects.get(id=pk)
-    if request.method =='POST':
-        data.slotsize-=1
-        data.save()
+    if 'email' in request.session:
+        user=User.objects.get(email=request.session['email'])
+        data=Slot.objects.get(id=pk)
+        if request.method =='POST':
+            data.slotsize-=1
+            data.save()
 
-        Appointment.objects.create(
+            Appointment.objects.create(
             name=request.POST['name'],
-            # user=user,
+            user=user,
             slot=data,
             phone=request.POST['phone'],
             weeks=request.POST['weeks'],
             timeslot=request.POST['timeslot'],
             description = request.POST['description'],
-        )
-        return render(request,'appointment-form.html',{'msg':'Appointment Send To Doctor Successfully!!','user':user,'data':data})
-    else:
-        return render(request,'appointment-form.html',{'user':user,'data':data})
+            )
+            return render(request,'appointment-form.html',{'msg':'Appointment Send To Doctor Successfully!!','user':user,'data':data})
+        else:
+            return render(request,'appointment-form.html',{'user':user,'data':data})
+    return redirect('login')
